angle based calculations

+++ commons/src/geometry.rs
@@ -4,6 +4,6 @@ pub fn distance(a: &[i32; 2], b: &[i32; 2]) -> f64 {
     (x_diff.powf(2.0) + y_diff.powf(2.0)).sqrt()
 }
 
-pub fn average(a: &[i32; 2], b: &[i32; 2]) -> [i32; 2] {
-    [(a[0] + b[0]) / 2, (a[1] + b[1]) / 2]
+pub fn average(a: &[f64; 2], b: &[f64; 2]) -> [f64; 2] {
+    [(a[0] + b[0]) / 2.0, (a[1] + b[1]) / 2.0]
 }
