+++ trigonometry/src/canvas_chart.rs
@@ -22,24 +22,33 @@ impl DrawWithState for CanvasChart {
     fn draw(&self, app_state: &AppState) -> Result<(), JsValue> {
         self.canvas.draw(app_state)?;
 
-        let chart_scale = self.canvas.offset[1] as f64;
+        let chart_scale = self.canvas.offset[1];
 
         self.canvas.context.draw_point_with_size_and_color(
-            &[((app_state.get_theta() * chart_scale).round()) as i32, (app_state.get_sin() * chart_scale) as i32],
-            2,
-            "red"
+            &[
+                (app_state.get_theta() * chart_scale).round(),
+                app_state.get_sin() * chart_scale,
+            ],
+            2.0,
+            "red",
         );
 
         self.canvas.context.draw_point_with_size_and_color(
-            &[((app_state.get_theta() * chart_scale).round()) as i32, (app_state.get_cos() * chart_scale) as i32],
-            2,
-            "blue"
+            &[
+                (app_state.get_theta() * chart_scale).round(),
+                app_state.get_cos() * chart_scale,
+            ],
+            2.0,
+            "blue",
         );
 
         self.canvas.context.draw_point_with_size_and_color(
-            &[((app_state.get_theta() * chart_scale).round()) as i32, (app_state.get_tan() * chart_scale) as i32],
-            2,
-            "magenta"
+            &[
+                (app_state.get_theta() * chart_scale).round(),
+                app_state.get_tan() * chart_scale,
+            ],
+            2.0,
+            "magenta",
         );
 
         Ok(())
